+++ docs/src/markdown/about/changelog.md
@@ -4,6 +4,11 @@
 
 - **NEW**: InlineHilite: Custom inline code block formatters can now be forced to raise an exception by raising a
   `InlineHiliteException`.
+- **NEW**: Snippets: Add new options to handle importing snippets from URL.
+- **NEW**: Snippets: Snippets will only swallow missing file errors (unless `check_paths` is enabled), all other errors
+  will be propagated up.
+- **NEW**: Snippets: When a file or URL is missing, raise `SnippetMissingError` instead of `IOError`.
+- **FIX**: Snippets: Small issues related to recursive inclusion of snippets.
 
 ## 9.4